Compile on windows
[git-p4: depot-paths = "//depot/qt/main/": change = 109893]

+++ src/kernel/qprocess_win.cpp
@@ -513,7 +513,7 @@ void QProcess::socketWrite( int )
     while ( !d->stdinBuf.isEmpty() && isRunning() ) {
 	if ( !WriteFile( d->pipeStdin[1],
 		    d->stdinBuf.head()->data() + d->stdinBufRead,
-		    QMIN( 8192, d->stdinBuf.head()->size() - d->stdinBufRead ),
+		    QMIN( (uint)8192, d->stdinBuf.head()->size() - d->stdinBufRead ),
 		    &written, 0 ) ) {
 	    d->lookup->start( 100 );
 	    return;
